requirements.txt
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.12.1
docutils==0.16
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
Markdown==3.3.3
Pillow==8.0.1
psycopg2==2.8.6
pycares==3.1.1
pycparser==2.20
pycurl==7.43.0.5
PyHamcrest==2.0.2
pytz==2020.1
requests==2.24.0
six==1.14.0
sqlparse==0.4.1
statistics==1.0.3.5
tornado==6.0.4
urllib3==1.25.11
whitenoise==5.2.0
zope.interface==5.1.0
